Modify : Media Chart Table Total Media &SCSS
