+++ src/routes/Dashboard/MediaAds/mediaAds.module.scss
@@ -1,3 +1,5 @@
+@use '/src/styles/constants/colors';
+
 .mediaAdsWrap {
   padding: 40px;
 
@@ -8,20 +10,66 @@
   }
 
   .content {
-    padding: 80px 40px 40px;
+    padding: 0 40px 40px;
     background-color: #ffffff;
     border-radius: 20px;
 
-    // .chartWrap {
+    .mediaAdChartWrap {
+      .chartInfo {
+        padding-right: 40px;
+        margin-top: -20px;
+        margin-right: 20px;
+        text-align: right;
 
-    // }
+        .chartChannelInfo {
+          position: relative;
+          display: inline-block;
+          padding-left: 15px;
+          margin-left: 40px;
+          font-size: 12px;
+          color: colors.$GRAY300;
+
+          &::after {
+            position: absolute;
+            top: 2px;
+            left: 0;
+            width: 8px;
+            height: 8px;
+            content: '';
+            // background-color: colors.$KAKAO;
+            border-radius: 4px;
+          }
+
+          &:nth-child(1)::after {
+            background-color: colors.$FACEBOOK;
+          }
+
+          &:nth-child(2)::after {
+            background-color: colors.$NAVER;
+          }
+
+          &:nth-child(3)::after {
+            background-color: colors.$GOOGLE;
+          }
+
+          &:nth-child(4)::after {
+            background-color: colors.$KAKAO;
+          }
+        }
+      }
+    }
+
     .tableWrap {
-      padding: 20px 0;
+      margin-top: 80px;
       overflow-x: auto;
       font-size: 12px;
 
       tr {
         border-bottom: 1px solid #edeff1;
+      }
+
+      .totalData {
+        color: colors.$Primary;
       }
 
       .table {
@@ -32,7 +80,7 @@
         .tableTitle {
           width: 300px;
           padding: 10px 0;
-          color: #94a2ad;
+          color: colors.$GRAY300;
           text-align: center;
         }
