+++ src/styles/constants/_colors.scss
@@ -18,6 +18,12 @@ $BLUE: #007aff;
 $INDIGO: #5856d6;
 $PURPLE: #af52de;
 $PINK: #ff2d55;
+$Primary: #586cf5;
+
+$GOOGLE: #ac8af8;
+$NAVER: #85da47;
+$FACEBOOK: #4fadf7;
+$KAKAO: #ffeb00;
 
 $GRAY50: #edeff1;
 $GRAY100: #d1d8dc;
